+++ robocorp-python-ls-core/src/robocorp_ls_core/protocols.py
@@ -243,7 +243,7 @@ class ILanguageServerClientBase(IRequestCancellable, Protocol):
         """
 
     def obtain_pattern_message_matcher(
-        self, message_pattern: Dict[str, str]
+        self, message_pattern: Dict[str, str], remove_on_match: bool = True
     ) -> IMessageMatcher:
         """
         Can be used as: